requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0
scipy>=1.8.0
pytz>=2022.1
python-dateutil>=2.8.0